+++ interface/js/editor.js
@@ -1,45 +1,9 @@
-import * as main from './main.js';
-import * as utils from './utilities.js';
-import Counter from './counter.js';
-import History from './history.js';
+import * as main from './screen.js';
 
 const HISTORY_LENGTH = 500;
 
-const screen = document.getElementById('screen');
-let counters;
-let history;
-
-const resetCounters = (historyLength) => {
-  try {
-    if (utils.mobileOrTabletCheck()) utils.requestFullscreen();
-
-    counters = {};
-    history = new History(historyLength);
-
-    const configElement = document.getElementById('config');
-    const config = JSON.parse(configElement.value);
-
-    screen.innerHTML = '';
-    screen.style.gridTemplate = main.generateGridTemplate(config.grid);
-    window.screenColor = config.color || 'white';
-    screen.style.setProperty('--screen-color', window.screenColor);
-
-    for (const counterID in config.counters) {
-      if (Object.hasOwn(config.counters, counterID)) {
-        counters[counterID] = new Counter(
-          screen,
-          counterID,
-          config.counters[counterID]
-        );
-      }
-    }
-  } catch (error) {
-    utils.log(error);
-  }
-};
-
 document.getElementById('increment').addEventListener('click', () => {
-  main.incrementAll(counters, history);
+  main.incrementAll();
 });
 document.addEventListener('keydown', (event) => {
   if (event.target.nodeName === 'BODY' && !event.ctrlKey && !event.altKey) {
@@ -47,22 +11,22 @@ document.addEventListener('keydown', (event) => {
       event.preventDefault();
       event.stopPropagation();
 
-      main.incrementAll(counters, history);
+      main.incrementAll();
     } else if (event.key === 'z') {
       event.preventDefault();
       event.stopPropagation();
 
-      main.undo(counters, history);
+      main.undo();
     } else if (event.key === 'r') {
       event.preventDefault();
       event.stopPropagation();
 
-      resetCounters(HISTORY_LENGTH);
+      main.resetCounters(HISTORY_LENGTH);
     }
   }
 });
 document.getElementById('reset').addEventListener('click', () => {
-  resetCounters(HISTORY_LENGTH);
+  main.resetCounters(HISTORY_LENGTH);
 });
 document.addEventListener('DOMContentLoaded', () => {
   document.getElementById('config').value = JSON.stringify(
@@ -70,5 +34,5 @@ document.addEventListener('DOMContentLoaded', () => {
     null,
     2 // eslint-disable-line no-magic-numbers
   );
-  resetCounters(HISTORY_LENGTH);
+  main.resetCounters(HISTORY_LENGTH);
 });